pyproject.toml
[project]
name = "housingpricecompetition"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=3.0.1,<4.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "scipy (>=1.17.0,<2.0.0)",
    "statsmodels (>=0.14.6,<0.15.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "xgboost (>=3.2.0,<4.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "optuna (>=4.7.0,<5.0.0)",
    "catboost (>=1.2.10,<2.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "notebook (>=7.5.3,<8.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "varclushi (>=0.1.0,<0.2.0)",
    "shap (>=0.50.0,<0.51.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
